Factored out header with pattern objects
